fix: resolve mangayomi client sync compatibility issues
Core changes:
- Create lib.rs exposing pub modules (app, db, globals, sync, user)
- Update main.rs to use library imports instead of local mod declarations

Bug fixes:
- Convert all i32 fields to i64 in sync models (history, manga, settings, update)
  This fixes deserialization errors with Dart int (64-bit) timestamps > i32::MAX
- Fix logout controller: Identity -> Option<Identity> to handle expired sessions

Debug improvements:
- Add raw payload logging to all sync controllers (settings, manga, history, updates)
- Return explicit JSON parse errors instead of silent 400s

+++ src/sync/history/model.rs
@@ -12,7 +12,7 @@ pub struct History {
     #[serde(rename = "chapterId")]
     pub chapter_id: i64,
     #[serde(rename = "itemType")]
-    pub item_type: i32,
+    pub item_type: i64,
     #[serde(skip_serializing)]
     pub user: Option<ObjectId>,
     #[serde(rename = "updatedAt")]
